backend/requirements.txt
# Core web framework
fastapi>=0.115.12
uvicorn>=0.34.3
python-multipart>=0.0.20
prometheus-fastapi-instrumentator>=6.1.0
prometheus-client>=0.22.1

# LangChain components
langchain>=0.3.25
langchain-core>=0.3.65
langchain-openai>=0.3.22
langgraph>=0.4.8
langgraph-checkpoint>=2.0.26
langgraph-checkpoint-postgres>=2.0.21
langsmith>=0.3.45

# OpenAI
openai>=1.86.0

# Data handling
pandas>=2.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0

# Authentication
bcrypt>=4.3.0
PyJWT>=2.10.1
email-validator>=2.2.0

# Environment setup
python-dotenv>=1.1.0
nest-asyncio>=1.6.0

# HTTP client
httpx>=0.28.1

# Database
supabase>=2.15.3
psycopg>=3.2.9
psycopg-pool>=3.2.6
asyncpg>=0.30.0

# Redis Cache
redis>=6.2.0
aioredis>=2.0.1
setuptools>=80.9.0

# MCP (Model Context Protocol)
langchain-mcp-adapters>=0.1.7
mcp>=1.9.4

# RAG (Retrieval-Augmented Generation)
pinecone>=7.0.2
langchain-pinecone>=0.2.8
